Add missing links
Now that we've renamed taxons with numbers in them, and regenerated
the taxon data json, there are some missing links.

Most of these are because we added some additional taxons in collections publisher,
so these need to be reflected in the script that fetches the taxons, for
the prototype to be able to render pages for them.

This is still a work in progress, as we are planning to automate the crawling
of the taxonomy.