Include rudimentary ordering of content by timestamp
Use the public_updated_at to apply a basic chronological order to
tagged content. Content items were previously rendered in the frontend
in whatever order they appear in the taxonomy-data.json. This allows
things like a view of the 'latest' content items to make at least some
kind of sense when viewed by the user.